CORE-6333 Added more tests for admin apps

+++ modules/belphegor-desktop/src/main/java/org/iplantc/de/admin/apps/client/views/toolbar/AdminAppsToolbarViewImpl.java
@@ -1,12 +1,12 @@
 package org.iplantc.de.admin.apps.client.views.toolbar;
 
+import org.iplantc.de.admin.apps.client.AdminAppsToolbarView;
 import org.iplantc.de.admin.apps.client.events.selection.AddCategorySelected;
 import org.iplantc.de.admin.apps.client.events.selection.CategorizeAppSelected;
 import org.iplantc.de.admin.apps.client.events.selection.DeleteCategorySelected;
 import org.iplantc.de.admin.apps.client.events.selection.MoveCategorySelected;
 import org.iplantc.de.admin.apps.client.events.selection.RenameCategorySelected;
 import org.iplantc.de.admin.apps.client.events.selection.RestoreAppSelected;
-import org.iplantc.de.admin.apps.client.AdminAppsToolbarView;
 import org.iplantc.de.apps.client.events.AppSearchResultLoadEvent;
 import org.iplantc.de.apps.client.events.selection.AppCategorySelectionChangedEvent;
 import org.iplantc.de.apps.client.events.selection.AppSelectionChangedEvent;
@@ -16,7 +16,9 @@ import org.iplantc.de.apps.client.presenter.toolBar.proxy.AppSearchRpcProxy;
 import org.iplantc.de.apps.client.views.toolBar.AppSearchField;
 import org.iplantc.de.client.models.apps.App;
 import org.iplantc.de.client.models.apps.AppCategory;
+import org.iplantc.de.commons.client.views.dialogs.IPlantPromptDialog;
 
+import com.google.common.base.Preconditions;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.uibinder.client.UiBinder;
@@ -32,8 +34,13 @@ import com.sencha.gxt.data.shared.loader.FilterPagingLoadConfig;
 import com.sencha.gxt.data.shared.loader.PagingLoadResult;
 import com.sencha.gxt.data.shared.loader.PagingLoader;
 import com.sencha.gxt.widget.core.client.Composite;
+import com.sencha.gxt.widget.core.client.Dialog;
+import com.sencha.gxt.widget.core.client.box.ConfirmMessageBox;
+import com.sencha.gxt.widget.core.client.box.PromptMessageBox;
 import com.sencha.gxt.widget.core.client.button.TextButton;
+import com.sencha.gxt.widget.core.client.event.DialogHideEvent;
 import com.sencha.gxt.widget.core.client.event.SelectEvent;
+import com.sencha.gxt.widget.core.client.form.TextField;
 import com.sencha.gxt.widget.core.client.toolbar.ToolBar;
 
 import java.util.List;
@@ -65,6 +72,8 @@ public class AdminAppsToolbarViewImpl extends Composite implements AdminAppsTool
     private List<AppCategory> appCategorySelection;
     private List<App> appSelection;
     private AppSearchRpcProxy proxy;
+
+    @Inject AdminAppsToolbarView.ToolbarAppearance appearance;
 
     @Inject
     AdminAppsToolbarViewImpl(@Assisted final PagingLoader<FilterPagingLoadConfig, PagingLoadResult<App>> loader) {
@@ -119,20 +128,26 @@ public class AdminAppsToolbarViewImpl extends Composite implements AdminAppsTool
     public void onAppCategorySelectionChanged(AppCategorySelectionChangedEvent event) {
         appCategorySelection = event.getAppCategorySelection();
 
-        boolean renameCategoryEnabled, deleteEnabled, moveCatEnabled;
+        boolean renameCategoryEnabled,
+            deleteEnabled,
+            moveCatEnabled,
+            addCategoryEnabled;
+
         switch (appCategorySelection.size()) {
             case 1:
                 renameCategoryEnabled = true;
                 deleteEnabled = true;
                 moveCatEnabled = true;
+                addCategoryEnabled = true;
                 break;
             default:
                 renameCategoryEnabled = false;
                 deleteEnabled = false;
                 moveCatEnabled = false;
+                addCategoryEnabled = true;
 
         }
-        addCategory.setEnabled(true);
+        addCategory.setEnabled(addCategoryEnabled);
         renameCategory.setEnabled(renameCategoryEnabled);
         deleteCat.setEnabled(deleteEnabled);
         moveCategory.setEnabled(moveCatEnabled);
@@ -166,21 +181,43 @@ public class AdminAppsToolbarViewImpl extends Composite implements AdminAppsTool
 
     @UiHandler("categorizeApp")
     void categorizeAppClicked(SelectEvent event) {
-//        presenter.onCategorizeAppClicked();
         fireEvent(new CategorizeAppSelected(appSelection));
     }
 
     @UiHandler("deleteApp")
     void deleteAppClicked(SelectEvent event) {
-//        presenter.onDeleteAppClicked();
-        fireEvent(new DeleteAppsSelected(appSelection));
+        ConfirmMessageBox msgBox = new ConfirmMessageBox(appearance.confirmDeleteAppWarning(),
+                                                         appearance.confirmDeleteAppTitle());
+
+        msgBox.addDialogHideHandler(new DialogHideEvent.DialogHideHandler() {
+            @Override
+            public void onDialogHide(DialogHideEvent event) {
+                if(!Dialog.PredefinedButton.YES.equals(event.getHideButton())){
+                    return;
+                }
+                fireEvent(new DeleteAppsSelected(appSelection));
+            }
+        });
+        msgBox.show();
     }
 
     @UiHandler("deleteCat")
     void deleteCatClicked(SelectEvent event) {
-//        presenter.onDeleteCatClicked();
-        // FIXME Move confirm dialog here
-        fireEvent(new DeleteCategorySelected(appCategorySelection));
+        Preconditions.checkState(appCategorySelection.size() == 1);
+        final AppCategory selectedAppCategory = appCategorySelection.iterator().next();
+        ConfirmMessageBox msgBox = new ConfirmMessageBox(appearance.confirmDeleteAppCategoryWarning(),
+                                                         appearance.confirmDeleteAppCategory(selectedAppCategory.getName()));
+        msgBox.addDialogHideHandler(new DialogHideEvent.DialogHideHandler() {
+            @Override
+            public void onDialogHide(DialogHideEvent event) {
+                if (!Dialog.PredefinedButton.YES.equals(event.getHideButton())) {
+                    return;
+                }
+
+                fireEvent(new DeleteCategorySelected(appCategorySelection));
+            }
+        });
+        msgBox.show();
     }
 
     @UiHandler("moveCategory")
@@ -192,39 +229,41 @@ public class AdminAppsToolbarViewImpl extends Composite implements AdminAppsTool
 
     @UiHandler("addCategory")
     void onAddCategoryClicked(SelectEvent event) {
-//        presenter.onAddAppCategoryClicked();
-        fireEvent(new AddCategorySelected(appCategorySelection));
+        final IPlantPromptDialog dlg = new IPlantPromptDialog(appearance.add(),
+                                                  0, "", null);
+        dlg.setHeadingText(appearance.addCategoryPrompt());
+        dlg.addOkButtonSelectHandler(new SelectEvent.SelectHandler() {
+            @Override
+            public void onSelect(SelectEvent event) {
+                fireEvent(new AddCategorySelected(appCategorySelection,
+                                                  dlg.getFieldText()));
+            }
+        });
+        dlg.show();
     }
-
-//    @Override
-//    public void init(final AdminAppsView.AdminPresenter presenter,
-//                     final AdminAppsView appView,
-//                     final HasAppSelectionChangedEventHandlers hasAppSelectionChangedEventHandlers,
-//                     final AppCategorySelectionChangedEvent.HasAppCategorySelectionChangedEventHandlers hasAppCategorySelectionChangedEventHandlers) {
-//        this.presenter = presenter;
-    // FIXME wire up search result load handler
-//        addAppSearchResultLoadEventHandler(appView);
-//        hasAppSelectionChangedEventHandlers.addAppSelectionChangedEventHandler(this);
-//        hasAppCategorySelectionChangedEventHandlers.addAppCategorySelectedEventHandler(this);
-//        proxy.setHasHandlers(asWidget());
-//        proxy.setMaskable(new IsMaskable() {
-//            @Override
-//            public void mask(String loadingMask) {
-//                appView.maskCenterPanel(loadingMask);
-//            }
-//
-//            @Override
-//            public void unmask() {
-//                appView.unMaskCenterPanel();
-//            }
-//        });
-//    }
 
     @UiHandler("renameCategory")
     void renameCategoryClicked(SelectEvent event) {
-//        presenter.onRenameAppCategoryClicked();
-        // FIXME Move dialog here
-        fireEvent(new RenameCategorySelected(appCategorySelection.iterator().next()));
+        Preconditions.checkState(appCategorySelection.size() == 1);
+        final AppCategory selectedAppCategory = appCategorySelection.iterator().next();
+         final PromptMessageBox msgBox = new PromptMessageBox(appearance.renameCategory(),
+                                                       appearance.renamePrompt());
+        final TextField field = ((TextField) msgBox.getField());
+        field.setAutoValidate(true);
+        field.setAllowBlank(false);
+        field.setText(selectedAppCategory.getName());
+        msgBox.addDialogHideHandler(new DialogHideEvent.DialogHideHandler() {
+            @Override
+            public void onDialogHide(DialogHideEvent event) {
+                if(!Dialog.PredefinedButton.OK.equals(event.getHideButton())){
+                    return;
+                }
+                fireEvent(new RenameCategorySelected(appCategorySelection.iterator().next(),
+                                                     field.getText()));
+            }
+        });
+
+        msgBox.show();
     }
 
     @UiHandler("restoreApp")
@@ -237,20 +276,4 @@ public class AdminAppsToolbarViewImpl extends Composite implements AdminAppsTool
     AppSearchField createAppSearchField() {
         return new AppSearchField(loader);
     }
-//
-//    @UiFactory
-//    PagingLoader<FilterPagingLoadConfig, PagingLoadResult<App>> createPagingLoader() {
-//        IplantDisplayStrings displayStrings = GWT.create(IplantDisplayStrings.class);
-//        // FIXME Fix this with injection
-//        AppsToolbarView.AppsToolbarAppearance appearance = GWT.create(AppsToolbarView.AppsToolbarAppearance.class);
-//        proxy = new AppSearchRpcProxy(appService, appSearchFactory, appFactory, appearance);
-//        PagingLoader<FilterPagingLoadConfig, PagingLoadResult<App>> loader = new PagingLoader<>(
-//                                                                                                   proxy);
-//
-//        AppLoadConfig appLoadConfig = appSearchFactory.loadConfig().as();
-//        loader.useLoadConfig(appLoadConfig);
-//
-//        return loader;
-//    }
-
 }
